+++ packages/wallets/evm-web3-wallets/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @thorswap-lib/evm-web3-wallets
 
+## 1.0.0-development.44
+
+### Patch Changes
+
+- Updated dependencies [[`0fb7dce`](https://github.com/thorswap/SwapKit/commit/0fb7dce1837816028e248b8f014a268a75a54190)]:
+  - @thorswap-lib/toolbox-evm@1.0.0-development.166
+  - @thorswap-lib/types@1.0.0-development.106
+
+## 1.0.0-development.43
+
+### Patch Changes
+
+- Updated dependencies [[`96b6212`](https://github.com/thorswap/SwapKit/commit/96b6212b57d3b66cdfc3fd014e9dc71fe8907a84)]:
+  - @thorswap-lib/types@1.0.0-development.105
+  - @thorswap-lib/toolbox-evm@1.0.0-development.165
+
 ## 1.0.0-development.42
 
 ### Patch Changes